requirements.txt
# Agent Studio Platform - Dependencies

# LangChain / LangGraph
langgraph>=1.0.6
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.1.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langsmith>=0.1.0
langfuse>=2.36.0

# Google Cloud
google-cloud-aiplatform>=1.50.0
google-genai>=1.0.0
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.14.0
google-cloud-discoveryengine>=0.13.0
google-cloud-resource-manager>=1.12.0

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
websockets>=12.0

# HTTP Client
httpx>=0.27.0
tenacity>=8.2.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Security (credential encryption at rest)
cryptography>=42.0.0

# Redis (shared state + caching)
redis[hiredis]>=5.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Dev Tools
ruff>=0.3.0
